Change duty cycle to integer
Change duty cycle value from decimal number to integer

+++ app/src/main/java/com/kynetics/android_pwm_example_app/PwmSettingsDialog.java
@@ -54,7 +54,7 @@ public class PwmSettingsDialog extends DialogFragment {
         /* Setup initial values */
         try {
             editTextFreq.setText(String.valueOf(mPwm.getFrequencyHz()));
-            editTextDutyCycle.setText(String.valueOf(mPwm.getDutyCycle()).trim());
+            editTextDutyCycle.setText(String.valueOf((int) mPwm.getDutyCycle()).trim());
             radioGroupPolarity.check(mPwm.getPolarity() == PwmPolarity.POLARITY_NORMAL ? R.id.radioButtonNormal : R.id.radioButtonInversed);
 
         } catch (IOException e) {
